Rework #filename tests for DRYness and robustness
- Use explicit subject w/ 1-liner `it` blocks
- Don't stub, use a real object.

In particular `original_filename` has enough complexity that it is
unclear that the behavior of stubbed objects corresponds to instances that
are actually possible to create in the wild.

**Case in point**, the nix directory test no longer passes, though it is
ostensibly doing the same thing!

```
expected: "_."
     got: "carrierwave"
```
That is a result of, effectively, `File.basename(File.expand_path(@file))`
between `original_filename` ane `path`.