New test, new try_uri method
The point is to avoid calling `File.expand_path` or `File.basename` on a thing that is not,
in fact, a local path: i.e., if it is a plausible URI that has a hostname.

In the future this might be improved to create and cache a URI object as
an attribute, clearing it when `file=` is called.

Fixes #2116